fix: move Rx extensions to shared module
