+++ shared/domain/build.gradle
@@ -29,10 +29,11 @@ android {
 }
 
 dependencies {
+    implementation project(':shared:components')
+
     implementation libraries.coreKtx
     implementation libraries.appcompat
     implementation libraries.material
-    implementation libraries.rxAndroid
     implementation libraries.rxJava
     implementation libraries.timber
 }
