+++ shared/domain/src/main/java/com/my/domain/usecase/FetchAllFavoriteUseCase.kt
@@ -2,7 +2,7 @@ package com.my.domain.usecase
 
 import com.my.domain.entity.MovieDetails
 import com.my.domain.repository.FavoriteRepository
-import com.my.domain.utils.applySchedulers
+import com.my.resources.extensions.applySchedulers
 import io.reactivex.Flowable
 
 class FetchAllFavoriteUseCase(private val repository: FavoriteRepository) {
