+++ shared/domain/src/main/java/com/my/domain/usecase/FetchCompositeMovieUseCase.kt
@@ -3,7 +3,7 @@ package com.my.domain.usecase
 import com.my.domain.entity.CompositeMovieEntity
 import com.my.domain.entity.Movie
 import com.my.domain.repository.MovieRepository
-import com.my.domain.utils.applySchedulers
+import com.my.resources.extensions.applySchedulers
 import io.reactivex.Flowable
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.observers.DisposableSingleObserver
