+++ shared/domain/src/main/java/com/my/domain/usecase/FetchMovieByIdUseCase.kt
@@ -1,7 +1,7 @@
 package com.my.domain.usecase
 
 import com.my.domain.repository.MovieRepository
-import com.my.domain.utils.applySchedulers
+import com.my.resources.extensions.applySchedulers
 
 class FetchMovieByIdUseCase(private val repository: MovieRepository) {
     operator fun invoke(id: String) = repository.fetchMovieByIdFromNetwork(id).applySchedulers()
