+++ shared/domain/src/main/java/com/my/domain/usecase/FetchPopularTvShowsUseCase.kt
@@ -2,7 +2,7 @@ package com.my.domain.usecase
 
 import com.my.domain.entity.TvShowsEntity
 import com.my.domain.repository.TvRepository
-import com.my.domain.utils.applySchedulers
+import com.my.resources.extensions.applySchedulers
 import io.reactivex.Single
 
 class FetchPopularTvShowsUseCase(private val repository: TvRepository) {
