+++ shared/domain/src/main/java/com/my/domain/usecase/IsFavoriteByIdUseCase.kt
@@ -1,7 +1,7 @@
 package com.my.domain.usecase
 
 import com.my.domain.repository.FavoriteRepository
-import com.my.domain.utils.applySchedulers
+import com.my.resources.extensions.applySchedulers
 
 class IsFavoriteByIdUseCase(private val repository: FavoriteRepository) {
     operator fun invoke(id: Int, isFavorite: (Boolean) -> Unit) =
